+++ src/Ui/Bundle/EventBundle/Handler/Unavailability/CreateForm.php
@@ -43,11 +43,11 @@ class CreateForm
         string $actionUrl,
         string $timezone
     ) {
-        $this->request     = $request;
-        $this->event       = $event;
-        $this->sheet       = $sheet;
-        $this->user        = $user;
-        $this->actionUrl   = $actionUrl;
+        $this->request = $request;
+        $this->event = $event;
+        $this->sheet = $sheet;
+        $this->user = $user;
+        $this->actionUrl = $actionUrl;
         $this->timezone = $timezone;
     }
 }